Endpoints

+++ CarsAndBids.Core/CQRS/Account/LoginViaRefreshTokenQuery.cs
@@ -9,7 +9,7 @@ namespace CarsAndBids.Core.CQRS.Account;
 
 public class LoginViaRefreshTokenQuery : IRequest<TokensDto>
 {
-    public string RefreshToken { get; set; }
+    public string? RefreshToken { get; set; }
 }
 
 public class LoginViaRefreshTokenQueryHandler(IJwtTokenService jwtTokenService,
@@ -17,38 +17,30 @@ public class LoginViaRefreshTokenQueryHandler(IJwtTokenService jwtTokenService,
 {
     public async Task<TokensDto> Handle(LoginViaRefreshTokenQuery request, CancellationToken cancellationToken)
     {
-        // Отримання токенів з бази за допомогою фільтра та включення користувача
         var refreshTokens = await repository.GetAsync(
             filter: token => token.Token == request.RefreshToken,
-            includeProperties: "User" // Включаємо навігаційну властивість User
+            includeProperties: "User"
         );
 
-        // Беремо перший токен
         var oldRefreshToken = refreshTokens.FirstOrDefault();
-        if (oldRefreshToken == null)
+        if (oldRefreshToken == null || oldRefreshToken.User == null)
         {
-            throw new Exception("Invalid refresh token");
+            throw new Exception("Invalid refresh token or user not found");
         }
 
-        // Перевірка, чи токен не прострочений
         if (oldRefreshToken.ExpiresOnUtc < DateTime.UtcNow)
         {
             throw new Exception("Refresh token has expired");
         }
 
-        // Генерація нового access-токена на основі даних користувача
         var accessToken = await jwtTokenService.CreateTokenAsync(oldRefreshToken.User);
-        
-        // Створення нового рефреш-токена
+
         var newRefreshToken = jwtTokenService.GenerateRefreshToken(oldRefreshToken.User);
 
-        // Додавання нового токена в базу
         await repository.InsertAsync(newRefreshToken);
 
-        // Видалення старого токена
         await repository.DeleteAsync(oldRefreshToken);
-        
-        // Повернення результату
+
         return new TokensDto
         {
             AccessToken = accessToken,
